+++ index2.js
@@ -1,0 +1,3 @@
+import { run } from './library.js';
+
+run();
